requirements.txt
Django==2.0.3
django-filter==1.1.0
djangorestframework==3.7.7
Markdown==2.6.11
numpy==1.14.2
pandas==0.22.0
python-dateutil==2.7.0
pytz==2018.3
six==1.11.0
